Rename gocql
