+++ libs/gocql/gocql/example_test.go
@@ -1,14 +1,14 @@
-package gocqltrace
+package gocql_test
 
 import (
 	"context"
+	gocqltrace "github.com/DataDog/dd-trace-go/libs/gocql/gocql"
 	"github.com/DataDog/dd-trace-go/tracer"
 	"github.com/gocql/gocql"
 )
 
 // To trace Cassandra commands, use our query wrapper TraceQuery.
 func Example() {
-
 	// Initialise a Cassandra session as usual, create a query.
 	cluster := gocql.NewCluster("127.0.0.1")
 	session, _ := cluster.CreateSession()
@@ -19,7 +19,7 @@ func Example() {
 	ctx := root.Context(context.Background())
 
 	// Wrap the query to trace it and pass the context for inheritance
-	tracedQuery := TraceQuery("ServiceName", tracer.DefaultTracer, query)
+	tracedQuery := gocqltrace.TraceQuery("ServiceName", tracer.DefaultTracer, query)
 	tracedQuery.WithContext(ctx)
 
 	// Execute your query as usual
