+++ libs/gocql/gocql/gocql_test.go
@@ -1,4 +1,4 @@
-package gocqltrace
+package gocql
 
 import (
 	"context"
